#123 check identity link count before trying to delete

+++ modules/flowable-engine/src/main/java/org/flowable/engine/impl/cmd/DeleteIdentityLinkCmd.java
@@ -16,6 +16,7 @@ package org.flowable.engine.impl.cmd;
 import org.flowable.engine.common.api.FlowableIllegalArgumentException;
 import org.flowable.engine.compatibility.Flowable5CompatibilityHandler;
 import org.flowable.engine.impl.interceptor.CommandContext;
+import org.flowable.engine.impl.persistence.CountingTaskEntity;
 import org.flowable.engine.impl.persistence.entity.TaskEntity;
 import org.flowable.engine.impl.util.Flowable5Util;
 import org.flowable.engine.task.IdentityLinkType;
@@ -81,7 +82,9 @@ public class DeleteIdentityLinkCmd extends NeedsActiveTaskCmd<Void> {
     } else if (IdentityLinkType.OWNER.equals(type)) {
       commandContext.getTaskEntityManager().changeTaskOwner(task, null);
     } else {
-      commandContext.getIdentityLinkEntityManager().deleteIdentityLink(task, userId, groupId, type);
+      if (((CountingTaskEntity) task).getIdentityLinkCount() != 0) {
+        commandContext.getIdentityLinkEntityManager().deleteIdentityLink(task, userId, groupId, type);
+      }     
     }
 
     commandContext.getHistoryManager().createIdentityLinkComment(taskId, userId, groupId, type, false);
